[BE] refactor: playkey, playvalue 클래스 분리  #763
