+++ backend/app/src/main/java/com/onair/hearit/app/playinghistory/infrastructure/buffer/PlayKey.java
@@ -1,0 +1,4 @@
+package com.onair.hearit.app.playinghistory.infrastructure.buffer;
+
+public record PlayKey(String userUuid, long hearitId) {
+    }
